requirements.txt
cffi==1.15.0
cryptography==36.0.2
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
